requirements.txt
editdistance==0.5.3
numpy==1.18.5
PyYAML==6.0
tensorflow==2.7.0
tqdm==4.41.1
